# apps/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "APTLY",
    "description": "Evidence-Grounded Multimodal AI Interview Coach \u2014 REST API\n\nPhase 0: Foundation scaffold. See /docs for OpenAPI specification.\nSee docs/api/contracts.md for planned endpoint contracts.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/jobs/analyze": {
      "post": {
        "tags": [
          "Jobs & Roles"
        ],
        "summary": "Analyze a Job Description",
        "description": "Parses a raw job description and extracts a structured RoleProfile.",
        "operationId": "analyze_job_description_api_v1_jobs_analyze_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/JobAnalyzeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/jobs/{job_id}": {
      "get": {
        "tags": [
          "Jobs & Roles"
        ],
        "summary": "Get Job & Role Profile",
        "description": "Fetches an existing job and its structured role profile.",
        "operationId": "get_job_api_v1_jobs__job_id__get",
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Job Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "JobAnalyzeRequest": {
        "properties": {
          "job_description": {
            "type": "string",
            "maxLength": 50000,
            "minLength": 20,
            "title": "Job Description",
            "description": "The raw text of the job description or posting.",
            "examples": [
              "Senior Backend Engineer with 5+ years of Python, FastAPI, PostgreSQL, and distributed systems experience."
            ]
          },
          "title": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 255
              },
              {
                "type": "null"
              }
            ],
            "title": "Title",
            "description": "Optional job title override."
          },
          "company": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 255
              },
              {
                "type": "null"
              }
            ],
            "title": "Company",
            "description": "Optional hiring company name."
          }
        },
        "type": "object",
        "required": [
          "job_description"
        ],
        "title": "JobAnalyzeRequest",
        "description": "Input payload for job description analysis."
      },
      "JobResponse": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          },
          "company": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Company"
          },
          "raw_text": {
            "type": "string",
            "title": "Raw Text"
          },
          "role_profile": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/RoleProfileResponse"
              },
              {
                "type": "null"
              }
            ]
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "title",
          "company",
          "raw_text",
          "created_at"
        ],
        "title": "JobResponse",
        "description": "Full Job entity with its associated RoleProfile."
      },
...
      "RoleProfileResponse": {
        "properties": {
          "schema_version": {
            "type": "string",
            "title": "Schema Version",
            "description": "Schema version. Increment when structure changes.",
            "default": "1.0",
            "examples": [
              "1.0",
              "1.1",
              "2.0"
            ]
          },
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "job_id": {
            "type": "string",
            "format": "uuid",
            "title": "Job Id"
          },
          "role_title": {
            "type": "string",
            "title": "Role Title"
          },
          "seniority": {
            "type": "string",
            "title": "Seniority"
          },
          "domain": {
            "type": "string",
            "title": "Domain"
          },
          "technical_skills": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Technical Skills"
          },
          "tools": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Tools"
          },
          "responsibilities": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Responsibilities"
          },
          "behavioral_competencies": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Behavioral Competencies"
          },
          "interview_topics": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Interview Topics"
          },
          "preferred_experience": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Preferred Experience"
          },
          "prompt_version": {
            "type": "string",
            "title": "Prompt Version",
            "default": "v1"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "job_id",
          "role_title",
          "seniority",
          "domain",
          "created_at"
        ],
        "title": "RoleProfileResponse",
        "description": "Structured role profile extracted from a job description."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}